--- Dictionary/src/main/java/Source/Dictionary.java
@@ -1,6 +1,0 @@
-package Source;
-
-import java.util.ArrayList;
-
-public class Dictionary extends ArrayList<Word> {
-}
