--- Dictionary/src/main/java/module-info.java
@@ -1,12 +1,0 @@
-module org.example.dictionary {
-    requires javafx.controls;
-    requires javafx.fxml;
-
-    requires org.controlsfx.controls;
-    requires com.dlsc.formsfx;
-    requires org.kordamp.bootstrapfx.core;
-    requires com.almasb.fxgl.all;
-
-    opens org.example.dictionary to javafx.fxml;
-//    exports org.example.dictionary;
-}
