--- Dictionary/src/main/resources/Texts/dictionaries.txt
@@ -1,2 +1,0 @@
-.love
-iu
