+++ Grasshopper-Li - Copy/.idea/encodings.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="Encoding">
+    <file url="file://$PROJECT_DIR$/Dictionary/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/Dictionary/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/FrontEnd/Eapp/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/FrontEnd/Eapp/src/main/resources" charset="UTF-8" />
+  </component>
+</project>
