+++ Grasshopper-Li - Copy/FrontEnd/Eapp/.idea/.gitignore
@@ -1,0 +1,3 @@
+# Default ignored files
+/shelf/
+/workspace.xml
