+++ Grasshopper-Li - Copy/FrontEnd/Eapp/data/RecentList.txt
@@ -1,0 +1,49 @@
+Bac dau boi tinh
+Bắc bán cầu
+Cao Bang
+Chu Van An
+Han luat
+Ho Chi Minh
+Bắc đẩu
+Cao Bằng
+Cham
+Ha Noi
+Hoàng Hoa Thám
+Hoà Bình
+A W O L
+2D image
+Dong Do
+Bien Hoa
+Blowing dust and wind
+Lac Long Quan
+hello
+hellos
+autobiographers
+gion toc gay
+Biên Hoà
+Binh Dinh
+Bắc đẩu bội tinh
+elaborateness
+elaborates
+elaborately
+-ment
+-trophin
+-mancy
+unassuming
+bac ban cau shshd
+bac ban cau
+bien hoa
+an com
+cham
+ba bó một giạ
+-manship
+ba bo mot gia
+ba bảy
+pho bo
+helloss
+-mer
+4th Generation Language (4GL)
+sidero-
+sidelong
+sideling
+Bình Định
