+++ Grasshopper-Li - Copy/FrontEnd/Eapp/src/main/java/BackEnd/Dictionary.java
@@ -1,0 +1,6 @@
+package BackEnd;
+
+import java.util.ArrayList;
+
+public class Dictionary extends ArrayList<Word> {
+}
