+++ Grasshopper-Li - Copy/FrontEnd/Eapp/src/main/java/module-info.java
@@ -1,0 +1,11 @@
+module org.example.eapp {
+    requires javafx.controls;
+    requires javafx.fxml;
+    requires javafx.web;
+
+    requires com.dlsc.formsfx;
+    requires com.almasb.fxgl.all;
+
+    opens org.example.eapp to javafx.fxml;
+    exports org.example.eapp;
+}
