+++ Grasshopper-Li - Copy/FrontEnd/Eapp/src/main/java/org/example/eapp/Word.java
@@ -1,0 +1,30 @@
+package org.example.eapp;
+
+public class Word {
+    private String word;
+    private String mean;
+
+    public Word(String word, String mean) {
+        this.word = word;
+        this.mean = mean;
+    }
+
+    public Word() {
+    }
+
+    public String getWord() {
+        return word;
+    }
+
+    public void setWord(String word) {
+        this.word = word;
+    }
+
+    public String getMean() {
+        return mean;
+    }
+
+    public void setMean(String mean) {
+        this.mean = mean;
+    }
+}
