+++ Grasshopper-Li - Copy/FrontEnd/Eapp/src/main/resources/org/example/eapp/main.fxml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.layout.AnchorPane?>
+
+<AnchorPane prefHeight="530.0" prefWidth="831.0" styleClass="root" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.example.eapp.MainController">
+   <children>
+      <AnchorPane fx:id="pane" layoutX="160.0" layoutY="65.0" prefHeight="400.0" prefWidth="600.0" />
+      <Button fx:id="translateBtn" layoutX="42.0" layoutY="208.0" mnemonicParsing="false" text="Translate" />
+      <Button fx:id="searchWordBtn" layoutX="47.0" layoutY="252.0" mnemonicParsing="false" text="Search" />
+   </children>
+</AnchorPane>
