+++ Grasshopper-Li - Copy/FrontEnd/Eapp/src/main/resources/org/example/eapp/translate.fxml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ComboBox?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.TextArea?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.example.eapp.Translator">
+   <children>
+      <TextArea fx:id="textTo" editable="false" layoutX="367.0" layoutY="138.0" maxWidth="-Infinity" minWidth="-Infinity" prefHeight="215.0" prefWidth="206.0" wrapText="true">
+      </TextArea>
+      <ComboBox fx:id="langFrom" layoutX="25.0" layoutY="92.0" onAction="#currentLanguageChanged" prefHeight="26.0" prefWidth="206.0" promptText="Language" styleClass="combobox">
+      </ComboBox>
+      <ComboBox fx:id="langTo" layoutX="367.0" layoutY="92.0" onAction="#targetLanguageChanged" prefHeight="26.0" prefWidth="206.0" promptText="Language" styleClass="combobox" />
+      <Button fx:id="transButton" layoutX="269.0" layoutY="225.0" mnemonicParsing="false" onAction="#onTransButtonClick" prefHeight="26.0" prefWidth="62.0" text="Dịch">
+      </Button>
+      <TextArea fx:id="textFrom" layoutX="25.0" layoutY="138.0" maxWidth="-Infinity" minWidth="-Infinity" prefHeight="215.0" prefWidth="206.0" wrapText="true">
+      </TextArea>
+      <Label layoutX="150.0" layoutY="29.0" prefHeight="50.0" prefWidth="300.0" text="TRANSLATION">
+         <font>
+            <Font name="Old English Text MT" size="33.0" />
+         </font>
+      </Label>
+   </children>
+</AnchorPane>
